player fluido en minimapa

+++ include/struct_map.h
@@ -6,7 +6,7 @@
 /*   By: mario <mario@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/18 17:33:05 by enrgil-p          #+#    #+#             */
-/*   Updated: 2026/05/09 13:21:26 by enrgil-p         ###   ########.fr       */
+/*   Updated: 2026/05/14 13:32:42 by mario            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -59,7 +59,7 @@ typedef struct s_map
 	int		ceiling_color[RGB_VALUES];
 	int		parse_checklist[CHECKLIST_SIZE];
 	char	**map;
-	int		player[POSITION];
+	double	player[POSITION];
 	char	spawn_orientation;
 	int		height;
 	int		width;
